requirements.txt
asgiref==3.3.1
Django==3.1.4
django-autoslug==1.9.8
pytz==2020.4
sqlparse==0.4.1
